Moved validation schema of CreatePayment form to it's own file
